+++ src/components/pod/pod.js
@@ -1,4 +1,3 @@
-import css from './../../utils/css';
 import React from 'react';
 import Icon from './../icon';
 import Link from './../link';
@@ -74,7 +73,10 @@ class Pod extends React.Component {
      * @property title
      * @type {String}
      */
-    title: React.PropTypes.string,
+    title: React.PropTypes.oneOfType([
+      React.PropTypes.string,
+      React.PropTypes.object
+    ]),
 
     /**
      * Aligns the title to left, right or center
@@ -130,6 +132,18 @@ class Pod extends React.Component {
    */
   componentWillMount = () => {
     this.setState({ collapsed: this.props.collapsed });
+  }
+
+  /**
+   * A lifecycle called immediatly before new props cause a re-render
+   * Resets the hover state if active
+   *
+   * @method componentWillReceiveProps
+   */
+  componentWillReceiveProps = () => {
+    if (this.state.hoverEdit) {
+      this.toggleHoverState(false);
+    }
   }
 
   /**
@@ -217,9 +231,12 @@ class Pod extends React.Component {
   }
 
   get mainClasses() {
-    return classNames("carbon-pod", this.props.className, css.clearfix, {
-      "carbon-pod--editable": this.props.onEdit
-    });
+    return classNames("carbon-pod", this.props.className,
+      `carbon-pod--${ this.props.alignTitle }`, {
+        "carbon-pod--editable": this.props.onEdit,
+        'carbon-pod--is-hovered': this.state.hoverEdit
+      }
+    );
   }
 
   /**
@@ -231,6 +248,7 @@ class Pod extends React.Component {
   get blockClasses() {
     return classNames(
       'carbon-pod__block',
+      `carbon-pod__block--padding-${this.props.padding}`,
       `carbon-pod__block--${this.props.as}`, {
         'carbon-pod__block--no-border': !this.props.border,
         'carbon-pod__block--footer': this.props.footer
@@ -327,7 +345,10 @@ class Pod extends React.Component {
   get edit() {
     if (!this.props.onEdit) { return null; }
 
-    let props = {};
+    let props = {
+      onMouseEnter: this.toggleHoverState.bind(this, true),
+      onMouseLeave: this.toggleHoverState.bind(this, false)
+    };
 
     if (typeof this.props.onEdit === "string") {
       props.to = this.props.onEdit;
@@ -340,6 +361,16 @@ class Pod extends React.Component {
     return (
       <Link icon="edit" className={ this.editActionClasses } { ...props } />
     );
+  }
+
+  /**
+   * Toggle the state of hovering the edit button.
+   *
+   * @method toggleHoverState
+   * @return {Void}
+   */
+  toggleHoverState = (val) => {
+    this.setState({ hoverEdit: val });
   }
 
   /**
@@ -357,8 +388,6 @@ class Pod extends React.Component {
 
     return (
       <div className={ this.mainClasses }>
-        { this.edit }
-
         <div className={ this.blockClasses } { ...props }>
           <div className={ this.contentClasses } >
             { this.podHeader }
@@ -366,6 +395,8 @@ class Pod extends React.Component {
           </div>
           { this.footer }
         </div>
+
+        { this.edit }
       </div>
     );
   }