+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 
 # 0.25.2
 
+## Bug Fix
+
+* Row now supports immutable children.
+* Row columns now clear when there are more columns than the defined number.
+* Editable Pod is now aligned properly with title.
+
 ## Minor Improvements
 
 * Added onTabClick props function to Tabs component to provide a way to add functionality which will only be called on tab click.